+++ src/codeactions/addFields/alAddXmlPortFieldsCodeCommand.ts
@@ -6,28 +6,34 @@ import { ALSyntaxWriter } from '../../allanguage/alSyntaxWriter';
 import { settings } from 'cluster';
 import { FieldsSelector } from './fieldsSelector';
 import { AZSymbolInformation } from '../../symbollibraries/azSymbolInformation';
+import { AZDocumentSymbolsLibrary } from '../../symbollibraries/azDocumentSymbolsLibrary';
 
 export class ALAddXmlPortFieldsCodeCommand extends ALBaseAddFieldsCodeCommand {
     elementType: string;
     commandTitle: string;
     
     constructor(context : DevToolsExtensionContext, newElementType: string, newCommandTitle: string) {
-        super(context, 'AZDevTools.ALAddXmlPort' + newElementType + 'CodeCommand');
+        super(context, 'AddXmlPort' + newElementType, 'AZDevTools.ALAddXmlPort' + newElementType + 'CodeCommand');
         this.elementType = newElementType;
         this.commandTitle = newCommandTitle;
     }
 
-    collectCodeActions(symbol: AZSymbolInformation, range: vscode.Range | vscode.Selection, actions: vscode.CodeAction[]) {
-        if ((symbol.kind == AZSymbolKind.XmlPortTableElement) ||
-            (symbol.kind == AZSymbolKind.XmlPortFieldElement) ||
-            (symbol.kind == AZSymbolKind.XmlPortFieldAttribute)) {
+    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[]) {
+        if ((symbol) &&         
+            ((symbol.kind == AZSymbolKind.XmlPortTableElement) ||
+             (symbol.kind == AZSymbolKind.XmlPortFieldElement) ||
+             (symbol.kind == AZSymbolKind.XmlPortFieldAttribute))) {
             let action = new vscode.CodeAction(this.commandTitle, vscode.CodeActionKind.QuickFix);
-            action.command = { command: this.name, title: this.commandTitle + '...' };
+            action.command = { 
+                command: this.name, 
+                title: this.commandTitle + '...',
+                arguments: [docSymbols, document, range]
+            };
             actions.push(action);
         }
     }
 
-    protected async runAsync(range: vscode.Range) {
+    protected async runAsync(docSymbols: AZDocumentSymbolsLibrary, document: vscode.TextDocument, range: vscode.Range) {
         //get required details from document source code
         let symbol = this._toolsExtensionContext.activeDocumentSymbols.findSymbolInRange(range);
         let isFieldSymbol = ((symbol.kind == AZSymbolKind.XmlPortFieldElement) || (symbol.kind == AZSymbolKind.XmlPortFieldAttribute));