+++ src/codeactions/sortSymbols/alSortPropertiesCommand.ts
@@ -2,48 +2,56 @@ import * as vscode from 'vscode';
 import { ALBaseSortCodeCommand } from "./alBaseSortCodeCommand";
 import { DevToolsExtensionContext } from "../../devToolsExtensionContext";
 import { AZSymbolInformation } from "../../symbollibraries/azSymbolInformation";
-import { ToolsDocumentSymbolsRequest } from '../../langserver/toolsDocumentSymbolsRequest';
-import { ToolsDocumentSymbolsResponse } from '../../langserver/toolsDocumentSymbolsResponse';
 import { AZSymbolKind } from '../../symbollibraries/azSymbolKind';
+import { AZDocumentSymbolsLibrary } from '../../symbollibraries/azDocumentSymbolsLibrary';
 
 export class ALSortPropertiesCommand extends ALBaseSortCodeCommand {
 
     constructor(context : DevToolsExtensionContext) {
-        super(context, 'AZDevTools.ALSortPropertiesCommand');
+        super(context, "SortProperties");
     }
 
-    collectCodeActions(symbol: AZSymbolInformation, range: vscode.Range | vscode.Selection, actions: vscode.CodeAction[]) {
-        if ((symbol.isALObject()) && (symbol.selectionRange.start.line == range.start.line)) {
-            let action = new vscode.CodeAction("Sort properties", vscode.CodeActionKind.QuickFix);
-            action.command = { command: this.name, title: 'Sort properties...' };
+    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[]) {
+        let edit: vscode.WorkspaceEdit | undefined = undefined;
+        let actionKind = vscode.CodeActionKind.QuickFix;
+
+        if (this.canRunOnSave(document.uri)) {
+            if ((context.only) && (context.only.contains(vscode.CodeActionKind.SourceFixAll))) {
+                actionKind = vscode.CodeActionKind.SourceFixAll;
+                let objList: AZSymbolInformation[] = [];        
+                docSymbols.findALObjectsInsideRange(range, objList);
+                for (let i=0; i<objList.length; i++)
+                    edit = this.prepareEdit(objList[i], document, edit);
+            }
+        } else {
+            //collect list of objects in selection range
+            if ((symbol) && (symbol.isALObject()) && (symbol.kind != AZSymbolKind.Interface) && (symbol.selectionRange.start.line == range.start.line))
+                edit = this.prepareEdit(symbol, document, undefined);
+        }
+
+        if (edit) {
+            let action = new vscode.CodeAction("Sort properties", actionKind);
+            action.edit = edit;
             actions.push(action);
         }
     }
 
-    protected async runAsync(range: vscode.Range) {
-        //load syntax tree
-        let document = vscode.window.activeTextEditor.document;
-        let source = document.getText();
+    protected prepareEdit(symbol: AZSymbolInformation, document: vscode.TextDocument, edit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit | undefined {
+        //collect PropertyList symbols
+        let propertyListSymbolsList: AZSymbolInformation[] = [];
+        symbol.collectChildSymbols(AZSymbolKind.PropertyList, true, propertyListSymbolsList);
+ 
+        if (propertyListSymbolsList.length == 0)
+            return edit;
 
-        //get document symbols with properties
-        let request : ToolsDocumentSymbolsRequest = new ToolsDocumentSymbolsRequest(source, document.uri.fsPath, true);
-        let response : ToolsDocumentSymbolsResponse | undefined = await this._toolsExtensionContext.toolsLangServerClient.getALDocumentSymbols(request);
-        if ((response) && (response.root)) {
-            let rootSymbol = AZSymbolInformation.fromAny(response.root);                    
-            rootSymbol.updateTree(true, true);
-
-            //collect PropertyList symbols
-            let propertyListSymbolsList: AZSymbolInformation[] = [];
-            rootSymbol.collectChildSymbols(AZSymbolKind.PropertyList, true, propertyListSymbolsList);
-            if (propertyListSymbolsList.length > 0) {
-                
-                await vscode.window.activeTextEditor.edit(editBuilder => {
-                    for (let i=0; i<propertyListSymbolsList.length; i++) {
-                        this.sortChildItems(propertyListSymbolsList[i], AZSymbolKind.Property, editBuilder);
-                    }
-                });
+        if (!edit)
+            edit = new vscode.WorkspaceEdit();
+        if (propertyListSymbolsList.length > 0) {
+            for (let i=0; i<propertyListSymbolsList.length; i++) {
+                this.sortChildItems(document.uri, propertyListSymbolsList[i], AZSymbolKind.Property, edit);
             }
         }
+        return edit;
     }
 
 }