+++ src/codeactions/sortSymbols/alSortReportColumnsCommand.ts
@@ -3,60 +3,64 @@ import { DevToolsExtensionContext } from "../../devToolsExtensionContext";
 import { ALBaseSortCodeCommand } from "./alBaseSortCodeCommand";
 import { AZSymbolKind } from '../../symbollibraries/azSymbolKind';
 import { AZSymbolInformation } from '../../symbollibraries/azSymbolInformation';
+import { AZDocumentSymbolsLibrary } from '../../symbollibraries/azDocumentSymbolsLibrary';
 
 export class ALSortReportColumnsCommand extends ALBaseSortCodeCommand {
 
     constructor(context : DevToolsExtensionContext) {
-        super(context, 'AZDevTools.ALSortReportColumnsCommand');
+        super(context, "SortReportColumns");
     }
 
-    collectCodeActions(symbol: AZSymbolInformation, range: vscode.Range | vscode.Selection, actions: vscode.CodeAction[]) {
-        if ((symbol.kind == AZSymbolKind.ReportDataItem) ||
-            (symbol.kind == AZSymbolKind.ReportColumn) ||
-            ((symbol.kind == AZSymbolKind.ReportObject) && (symbol.selectionRange.start.line == range.start.line))) {
-            let action = new vscode.CodeAction("Sort data item columns", vscode.CodeActionKind.QuickFix);
-            action.command = { command: this.name, title: 'Sort data item columns...' };
+    collectCodeActions(docSymbols: AZDocumentSymbolsLibrary, symbol: AZSymbolInformation, document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, actions: vscode.CodeAction[]) {
+        let edit: vscode.WorkspaceEdit | undefined = undefined;
+        let actionKind = vscode.CodeActionKind.QuickFix;        
+        
+        if (this.canRunOnSave(document.uri)) {
+            if ((context.only) && (context.only.contains(vscode.CodeActionKind.SourceFixAll))) {
+                actionKind = vscode.CodeActionKind.SourceFixAll;
+                let objList: AZSymbolInformation[] = [];        
+                docSymbols.findSymbolsInsideRange(range, AZSymbolKind.ReportObject, objList);
+                for (let i=0; i<objList.length; i++)
+                    edit = this.prepareEdit(objList[i], document, edit);
+            }
+        } else {
+            //prepare code fix
+            if ((symbol) &&
+                ((symbol.kind == AZSymbolKind.ReportDataItem) ||
+                (symbol.kind == AZSymbolKind.ReportColumn) ||
+                ((symbol.kind == AZSymbolKind.ReportObject) && (symbol.selectionRange.start.line == range.start.line))))
+                edit = this.prepareEdit(symbol, document, edit);
+        }
+
+        if (edit) {
+            let action = new vscode.CodeAction("Sort data item columns", actionKind);
+            action.edit = edit; 
             actions.push(action);
         }
     }
 
-    protected async runAsync(range: vscode.Range) {
-        // Get required details from document source code
-        let symbol = this._toolsExtensionContext.activeDocumentSymbols.findSymbolInRange(range);
-        if (!symbol) {
-            return;
-        }
-        
+    protected prepareEdit(symbol: AZSymbolInformation, document: vscode.TextDocument, edit: vscode.WorkspaceEdit | undefined): vscode.WorkspaceEdit {
+        if (!edit)
+            edit = new vscode.WorkspaceEdit();
         switch (symbol.kind) {
             case AZSymbolKind.ReportDataItem:
-                await this.sortDataItemAsync(symbol);
+                this.sortChildItems(document.uri, symbol, AZSymbolKind.ReportColumn, edit);
                 break;
             case AZSymbolKind.ReportColumn:
                 let dataItemSymbol = symbol.findParentByKind(AZSymbolKind.ReportDataItem);
                 if (dataItemSymbol)
-                    await this.sortDataItemAsync(symbol);
+                    this.sortChildItems(document.uri, dataItemSymbol, AZSymbolKind.ReportColumn, edit);
                 break;
             case AZSymbolKind.ReportObject:
                 let dataItemSymbolsList: AZSymbolInformation[] = [];
                 symbol.collectChildSymbols(AZSymbolKind.ReportDataItem, true, dataItemSymbolsList);
                 if (dataItemSymbolsList.length > 0)
-                    await this.sortDataItemsListAsync(dataItemSymbolsList);
+                    for (let i=0; i<dataItemSymbolsList.length; i++) {
+                        this.sortChildItems(document.uri, dataItemSymbolsList[i], AZSymbolKind.ReportColumn, edit);
+                    }
                 break;
         }
-    }
-
-    protected async sortDataItemsListAsync(dataItemSymbolsList: AZSymbolInformation[]) {
-        await vscode.window.activeTextEditor.edit(editBuilder => {
-            for (let i=0; i<dataItemSymbolsList.length; i++) {
-                this.sortChildItems(dataItemSymbolsList[i], AZSymbolKind.ReportColumn, editBuilder);
-            }
-        });
-    }
-
-    protected async sortDataItemAsync(dataItemSymbol: AZSymbolInformation) {
-        await vscode.window.activeTextEditor.edit(editBuilder => {
-            this.sortChildItems(dataItemSymbol, AZSymbolKind.ReportColumn, editBuilder);
-        });
+        return edit;
     }
 
 }