+++ src/objectwizards/wizards/alInterfaceWizardData.ts
@@ -1,0 +1,12 @@
+export class ALInterfaceWizardData {
+    objectName : string;
+    baseCodeunitName: string;
+    codeunitList: string[] | undefined;
+
+    constructor() {
+        this.objectName = '';
+        this.baseCodeunitName = '';
+        this.codeunitList = undefined;
+    }
+
+}